No collections, only entities and entries
