+++ spire/journal/version.py
@@ -3,4 +3,3 @@ Journal and collection modules of Spire library and API version.
 """
 
 SPIRE_JOURNALS_VERSION = "0.1.2"
-SPIRE_COLLECTIONS_VERSION = "0.0.7"
